refactor(rust cycle 17): split mailbox/mime_utils 448->187 + mime_body(132) + mime_attachments(139)
