+++ src/bin/email_api_dir/admin_ops/mod.rs
@@ -234,12 +234,14 @@ pub(crate) struct AdminUsersQuery {
 
 pub mod user_handlers;
 pub mod cr_handlers;
+pub mod cr_patch_helpers;
 pub mod ai_handlers;
 pub mod hermes_handlers;
 pub mod diag_handlers;
 
 pub use user_handlers::*;
 pub use cr_handlers::*;
+pub use cr_patch_helpers::*;
 pub use ai_handlers::*;
 pub use hermes_handlers::*;
 pub use diag_handlers::*;
