+++ src/bin/email_api_dir/mailbox/mime_utils.rs
@@ -1,10 +1,14 @@
-// MIME decoding / attachment extraction / EmailDto helpers extraits de mailbox/mod.rs
+// MIME DTO + address parsing helpers (post-Sprint 17 split).
+// Body decoding lives in mime_body.rs; attachment extraction in mime_attachments.rs.
 use super::super::*;
 #[allow(unused_imports)]
 use base64::Engine as _;
 use chrono::Utc;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
+
+use super::mime_attachments::extract_attachments_for_ui;
+use super::mime_body::decoded_mail_body_for_ui;
 
 #[derive(Serialize, Deserialize)]
 pub(crate) struct EmailAddressDto {
@@ -76,271 +80,6 @@ pub(crate) fn strip_tags(html: &str) -> String {
         }
     }
     out.split_whitespace().collect::<Vec<_>>().join(" ")
-}
-
-fn compact_preview(text: &str) -> String {
-    text.split_whitespace().collect::<Vec<_>>().join(" ")
-}
-
-fn raw_mime_from_email(email: &Email) -> String {
-    if email.headers.is_empty() {
-        email.body.clone()
-    } else {
-        let headers = email
-            .headers
-            .iter()
-            .map(|(k, v)| format!("{}: {}", k, v))
-            .collect::<Vec<_>>()
-            .join("\r\n");
-        format!("{}\r\n\r\n{}", headers, email.body)
-    }
-}
-
-fn walk_mime_parts(part: &mailparse::ParsedMail<'_>, html: &mut Option<String>, text: &mut Option<String>) {
-    if part.subparts.is_empty() {
-        let mt = part.ctype.mimetype.to_ascii_lowercase();
-        if mt == "text/html" && html.is_none() {
-            if let Ok(b) = part.get_body() {
-                *html = Some(b);
-            }
-        } else if mt == "text/plain" && text.is_none() {
-            if let Ok(b) = part.get_body() {
-                *text = Some(b);
-            }
-        }
-        return;
-    }
-
-    for sub in &part.subparts {
-        walk_mime_parts(sub, html, text);
-    }
-}
-
-fn decode_parts_from_parsed(parsed: &mailparse::ParsedMail<'_>) -> Option<(String, String, String)> {
-    let mut html = None;
-    let mut text = None;
-    walk_mime_parts(parsed, &mut html, &mut text);
-
-    if let Some(html_body) = html {
-        let preview = compact_preview(&strip_tags(&html_body)).chars().take(160).collect();
-        return Some((html_body, "html".to_string(), preview));
-    }
-    if let Some(text_body) = text {
-        let preview = compact_preview(&text_body).chars().take(160).collect();
-        return Some((text_body, "text".to_string(), preview));
-    }
-
-    None
-}
-
-fn looks_like_raw_multipart_dump(body: &str) -> bool {
-    let b = body.trim_start();
-    b.starts_with("--") && b.contains("Content-Type:")
-}
-
-fn decode_from_synthetic_boundary(body: &str) -> Option<(String, String, String)> {
-    let first_line = body.lines().next()?.trim();
-    if !first_line.starts_with("--") {
-        return None;
-    }
-    let boundary = first_line
-        .trim_start_matches("--")
-        .trim_end_matches("--")
-        .trim();
-    if boundary.is_empty() {
-        return None;
-    }
-
-    let synthetic = format!(
-        "Content-Type: multipart/alternative; boundary=\"{}\"\r\n\r\n{}",
-        boundary, body
-    );
-    let parsed = mailparse::parse_mail(synthetic.as_bytes()).ok()?;
-    decode_parts_from_parsed(&parsed)
-}
-
-fn infer_attachment_kind(content_type: &str, filename: &str) -> &'static str {
-    let ct = content_type.to_ascii_lowercase();
-    let lower_name = filename.to_ascii_lowercase();
-    if ct.starts_with("image/") {
-        "image"
-    } else if ct == "application/pdf" || lower_name.ends_with(".pdf") {
-        "pdf"
-    } else if ct.contains("word")
-        || lower_name.ends_with(".doc")
-        || lower_name.ends_with(".docx")
-        || lower_name.ends_with(".odt")
-    {
-        "doc"
-    } else if ct.contains("sheet")
-        || lower_name.ends_with(".xls")
-        || lower_name.ends_with(".xlsx")
-        || lower_name.ends_with(".csv")
-    {
-        "spreadsheet"
-    } else if ct.contains("presentation")
-        || lower_name.ends_with(".ppt")
-        || lower_name.ends_with(".pptx")
-    {
-        "presentation"
-    } else if ct.starts_with("audio/") {
-        "audio"
-    } else if ct.starts_with("video/") {
-        "video"
-    } else if ct.contains("zip")
-        || ct.contains("gzip")
-        || ct.contains("tar")
-        || ct.contains("7z")
-        || lower_name.ends_with(".zip")
-        || lower_name.ends_with(".tar")
-        || lower_name.ends_with(".gz")
-        || lower_name.ends_with(".7z")
-    {
-        "archive"
-    } else {
-        "other"
-    }
-}
-
-#[derive(Clone)]
-pub(crate) struct ExtractedAttachment {
-    pub id: String,
-    pub filename: String,
-    pub content_type: String,
-    pub size: u64,
-    pub kind: String,
-    pub data: Vec<u8>,
-}
-
-fn walk_mime_attachments(
-    part: &mailparse::ParsedMail<'_>,
-    out: &mut Vec<ExtractedAttachment>,
-    index: &mut usize,
-) {
-    if !part.subparts.is_empty() {
-        for sub in &part.subparts {
-            walk_mime_attachments(sub, out, index);
-        }
-        return;
-    }
-
-    let content_type = part.ctype.mimetype.to_ascii_lowercase();
-    let disp = part.get_content_disposition();
-    let disp_kind = format!("{:?}", disp.disposition).to_ascii_lowercase();
-    let filename = disp
-        .params
-        .get("filename")
-        .cloned()
-        .or_else(|| part.ctype.params.get("name").cloned())
-        .filter(|s| !s.trim().is_empty())
-        .unwrap_or_else(|| format!("attachment-{}", *index + 1));
-
-    let is_attachment = disp_kind == "attachment"
-        || disp.params.contains_key("filename")
-        || part.ctype.params.contains_key("name")
-        || (!content_type.starts_with("text/") && content_type != "application/pgp-signature");
-
-    if !is_attachment {
-        return;
-    }
-
-    let bytes = part.get_body_raw().unwrap_or_default();
-    let kind = infer_attachment_kind(&content_type, &filename).to_string();
-    let id = format!("att-{}", *index);
-    *index += 1;
-
-    out.push(ExtractedAttachment {
-        id,
-        filename,
-        content_type,
-        size: bytes.len() as u64,
-        kind,
-        data: bytes,
-    });
-}
-
-pub(crate) fn extract_attachments_for_ui(email: &Email) -> Vec<ExtractedAttachment> {
-    let raw_mime = raw_mime_from_email(email);
-    if let Ok(parsed) = mailparse::parse_mail(raw_mime.as_bytes()) {
-        let mut out = Vec::new();
-        let mut idx = 0usize;
-        walk_mime_attachments(&parsed, &mut out, &mut idx);
-        return out;
-    }
-
-    if looks_like_raw_multipart_dump(&email.body) {
-        let first_line = match email.body.lines().next() {
-            Some(l) => l.trim(),
-            None => return Vec::new(),
-        };
-        let boundary = first_line
-            .trim_start_matches("--")
-            .trim_end_matches("--")
-            .trim();
-        if boundary.is_empty() {
-            return Vec::new();
-        }
-        let synthetic = format!(
-            "Content-Type: multipart/mixed; boundary=\"{}\"\r\n\r\n{}",
-            boundary, email.body
-        );
-        if let Ok(parsed) = mailparse::parse_mail(synthetic.as_bytes()) {
-            let mut out = Vec::new();
-            let mut idx = 0usize;
-            walk_mime_attachments(&parsed, &mut out, &mut idx);
-            return out;
-        }
-    }
-
-    Vec::new()
-}
-
-fn decoded_mail_body_for_ui(email: &Email) -> (String, String, String) {
-    let raw_mime = raw_mime_from_email(email);
-
-    if let Ok(parsed) = mailparse::parse_mail(raw_mime.as_bytes()) {
-        if let Some(decoded) = decode_parts_from_parsed(&parsed) {
-            return decoded;
-        }
-
-        if let Ok(body) = parsed.get_body() {
-            if looks_like_raw_multipart_dump(&body) {
-                if let Some(decoded) = decode_from_synthetic_boundary(&body) {
-                    return decoded;
-                }
-            }
-
-            let looks_html = body.to_ascii_lowercase().contains("<html")
-                || body.contains("</")
-                || body.contains("<p");
-            if looks_html {
-                let preview = compact_preview(&strip_tags(&body)).chars().take(160).collect();
-                return (body, "html".to_string(), preview);
-            }
-            let preview = compact_preview(&body).chars().take(160).collect();
-            return (body, "text".to_string(), preview);
-        }
-    }
-
-    if let Some(decoded) = decode_from_synthetic_boundary(&email.body) {
-        return decoded;
-    }
-
-    let body_type = if email.body.to_ascii_lowercase().contains("<html")
-        || email.body.contains("</")
-        || email.body.contains("<p")
-    {
-        "html"
-    } else {
-        "text"
-    };
-    let plain = if body_type == "html" {
-        strip_tags(&email.body)
-    } else {
-        email.body.clone()
-    };
-    let preview = compact_preview(&plain).chars().take(160).collect();
-    (email.body.clone(), body_type.to_string(), preview)
 }
 
 pub(crate) fn email_to_dto(email: &Email, folder: &str, include_body: bool) -> EmailDto {
