+++ src/bin/email_api_dir/mailbox/mod.rs
@@ -8,6 +8,8 @@ pub mod drafts_handlers;
 
 mod folder_utils;
 mod mime_utils;
+mod mime_body;
+mod mime_attachments;
 
 // Le mod.rs originel exposait via glob les types du parent (Arc, Logic, Email,
 // Responder, bson, monitoring, ...). Après split minimal (helpers déplacés),
@@ -26,6 +28,6 @@ pub(crate) use folder_utils::{
     canonical_folder, folder_to_mailboxes, resolve_user_id, EmailListQuery,
 };
 pub(crate) use mime_utils::{
-    email_to_dto, extract_attachments_for_ui, parse_address, strip_tags, EmailAddressDto, EmailDto,
-    ExtractedAttachment,
+    email_to_dto, parse_address, strip_tags, EmailAddressDto, EmailDto,
 };
+pub(crate) use mime_attachments::{extract_attachments_for_ui, ExtractedAttachment};
